Refactor llm_call out of actions.llm.utils and stop rail actions loading Colang at import-time. Decouples actions from Colang
