+++ nemoguardrails/library/content_safety/actions.py
@@ -17,7 +17,6 @@ import logging
 from typing import Any, Dict, FrozenSet, Optional, cast
 
 from nemoguardrails.actions.actions import action
-from nemoguardrails.actions.llm.utils import llm_call, warn_if_truncated
 from nemoguardrails.actions.rail_outcome import RailOutcome
 from nemoguardrails.context import llm_call_info_var
 from nemoguardrails.llm.cache import CacheInterface
@@ -28,6 +27,7 @@ from nemoguardrails.llm.cache.utils import (
     extract_llm_stats_for_cache,
     get_from_cache_and_restore_stats,
 )
+from nemoguardrails.llm.call import llm_call, warn_if_truncated
 from nemoguardrails.llm.taskmanager import LLMTaskManager
 from nemoguardrails.logging.explain import LLMCallInfo
 from nemoguardrails.types import LLMModel
