+++ nemoguardrails/library/llama_guard/actions.py
@@ -18,9 +18,9 @@ from collections.abc import Mapping
 from typing import List, Optional, Tuple
 
 from nemoguardrails.actions import action
-from nemoguardrails.actions.llm.utils import llm_call
 from nemoguardrails.actions.rail_outcome import RailOutcome
 from nemoguardrails.context import llm_call_info_var
+from nemoguardrails.llm.call import llm_call
 from nemoguardrails.llm.taskmanager import LLMTaskManager
 from nemoguardrails.llm.types import Task
 from nemoguardrails.logging.explain import LLMCallInfo
