+++ nemoguardrails/library/patronusai/actions.py
@@ -20,10 +20,10 @@ from collections.abc import Mapping
 from typing import List, Literal, Optional, Tuple, Union
 
 from nemoguardrails.actions import action
-from nemoguardrails.actions.llm.utils import llm_call
 from nemoguardrails.actions.rail_outcome import RailOutcome
 from nemoguardrails.context import llm_call_info_var
 from nemoguardrails.http import HTTPClient, http_call
+from nemoguardrails.llm.call import llm_call
 from nemoguardrails.llm.taskmanager import LLMTaskManager
 from nemoguardrails.llm.types import Task
 from nemoguardrails.logging.explain import LLMCallInfo
