+++ nemoguardrails/library/self_check/utils.py
@@ -18,11 +18,11 @@ import warnings
 from collections.abc import Collection
 from typing import Any, Dict, List, Optional
 
-from nemoguardrails.actions.llm.utils import llm_call, warn_if_truncated
-from nemoguardrails.colang.v1_0.runtime.flows import _get_flow_params, _normalize_flow_id
 from nemoguardrails.context import llm_call_info_var
+from nemoguardrails.llm.call import llm_call, warn_if_truncated
 from nemoguardrails.logging.explain import LLMCallInfo
 from nemoguardrails.types import LLMModel
+from nemoguardrails.utils import _get_flow_params, _normalize_flow_id
 
 log = logging.getLogger(__name__)
